Post to logs when in dispute
