+++ index.js
@@ -30,6 +30,7 @@ const {
   addReport,
   updateShinyStatuses,
 } = require('./other/shinySquad.js');
+const Votes = require('./other/votes.js');
 const { extractMessageDate } = require('./other/ocr.js');
 
 const client = new Discord.Client({
@@ -70,24 +71,7 @@ for (const file of buttonCommandsFiles) {
   client.buttonCommands.set(command.name, command);
 }
 
-class Votes { 
-  #votes_verified = {};
-  #votes_denied = {};
-  hasVoted = (messageID, userID) => {
-      return this.#votes_verified[messageID]?.has(userID) || this.#votes_denied[messageID]?.has(userID);
-  }
-  verified = (messageID) => {
-      return this.#votes_verified[messageID] = this.#votes_verified[messageID] ?? new Set();
-  };
-  denied = (messageID) => {
-      return this.#votes_denied[messageID] = this.#votes_denied[messageID] ?? new Set();
-  };
-  delete = (messageID) => {
-      return delete this.#votes_verified[messageID] && delete this.#votes_denied[messageID];
-  };
-};
-
-client.votes = new Votes();;
+client.votes = new Votes(client);
 const cooldowns = new Discord.Collection();
 
 const cooldownTimeLeft = (type, seconds, userID) => {
